feat: refactor plugin operation job handling by separating start logic into core.start.ts
